app/src/main/java/dev/nuris/footballleague/module/presenter/TeamPresenter.kt
package dev.nuris.footballleague.module.presenter

import com.google.gson.Gson
import dev.nuris.footballleague.api.ApiRepository
import dev.nuris.footballleague.api.TheSportDBApi
import dev.nuris.footballleague.helper.CoroutineContextProvider
import dev.nuris.footballleague.model.StandingResponse
import dev.nuris.footballleague.model.TeamResponse
import dev.nuris.footballleague.module.view.StandingView
import dev.nuris.footballleague.module.view.TeamView
import kotlinx.coroutines.GlobalScope
import kotlinx.coroutines.launch

class TeamPresenter(private val view: TeamView,
                    private val apiRepository: ApiRepository,
                    private val gson: Gson,
                    private val context: CoroutineContextProvider = CoroutineContextProvider()) {

    fun getTeamList(leagueId: String) {
        view.showLoading()
        GlobalScope.launch(context.main) {
            val data = gson.fromJson(apiRepository.doRequestAsync(TheSportDBApi.getTeam(leagueId)).await(),
                TeamResponse::class.java)
            view.showTeamList(data)
            view.hideLoading()
        }
    }
}